Validation & API Fixes

+++ src/controllers/student/courseController.ts
@@ -5,6 +5,7 @@ import { sendResponse, sendError } from '../../utils/commonResponse';
 import runTransaction from '../../utils/runTransaction';
 import { ILike } from 'typeorm';
 import { createActivityLog } from '../../utils/activityLog';
+import { Semester } from '../../entity/Semester';
 
 // Create a new course
 export const createCourse = async (req: Request, res: Response) => {
@@ -51,7 +52,7 @@ export const createCourse = async (req: Request, res: Response) => {
 // Get all courses
 export const getAllCourses = async (req: Request, res: Response) => {
   try {
-    const { page, limit, name } = req.query;
+    const { page, limit, name, is_dropdown } = req.query;
 
     // Convert page and limit parameters to numbers
     const pageNumber = page ? parseInt(page as string) : 1;
@@ -68,15 +69,27 @@ export const getAllCourses = async (req: Request, res: Response) => {
       filter.name = ILike(`%${name}%`);
     }
 
-    // Fetch courses with pagination and filter
-    const [courses, totalCount] = await AppDataSource.getRepository(
-      Course,
-    ).findAndCount({
-      where: filter,
-      order: { createdAt: 'DESC' },
-      skip: offset,
-      take: limitNumber,
-    });
+    let courses;
+    let totalCount;
+    if (is_dropdown) {
+      [courses, totalCount] = await AppDataSource.getRepository(
+        Course,
+      ).findAndCount({
+        where: filter,
+        order: { name: 'ASC' },
+        skip: offset,
+        take: limitNumber,
+      });
+    } else {
+      [courses, totalCount] = await AppDataSource.getRepository(
+        Course,
+      ).findAndCount({
+        where: filter,
+        order: { createdAt: 'DESC' },
+        skip: offset,
+        take: limitNumber,
+      });
+    }
 
     const totalNoOfRecords = courses.length;
 
@@ -158,6 +171,7 @@ export const deleteCourseById = async (req: Request, res: Response) => {
     const queryRunner = AppDataSource.createQueryRunner();
     await runTransaction(queryRunner, async () => {
       const courseRepository = queryRunner.manager.getRepository(Course);
+      const semesterRepository = queryRunner.manager.getRepository(Semester);
       const course = await courseRepository.findOne({
         where: { id: +id },
       });
@@ -165,6 +179,20 @@ export const deleteCourseById = async (req: Request, res: Response) => {
         sendError(res, 404, 'Course not found');
         return;
       }
+
+      const semester = await semesterRepository.findOne({
+        where: { course: { id: +id } },
+      });
+
+      if (semester) {
+        sendError(
+          res,
+          400,
+          'A Semester(s) already exists in this Course so it cannot be deleted',
+        );
+        return;
+      }
+
       await courseRepository.remove(course);
 
       await createActivityLog(
